% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRToAnndata_2024.R
\name{ArchrToAnndata}
\alias{ArchrToAnndata}
\title{ArchRToAnndata}
\usage{
ArchrToAnndata(
  object,
  outpath,
  assays,
  assay.types,
  markersDF = NULL,
  groups = NULL,
  reductions = NULL,
  export_links = FALSE,
  export_pwm = FALSE
)
}
\arguments{
\item{object}{ArchR object}

\item{outpath}{output dir}

\item{assays}{vector of assay names to export}

\item{assay.types}{vector of the type of assays, mainly peak, gene, motif}

\item{markersDF}{a named list of marker df vars}

\item{groups}{vector of cell metadata colnames to export}

\item{reductions}{vector of reduction names to export}

\item{export_links}{whether to expport coaccess link informs}

\item{export_pwm}{whether to export pwm informas}
}
\description{
ArchRToAnndata
}
