% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GiottoToAnndata_2024.R
\name{Function getNoshiftSpatialLocs}
\alias{Function getNoshiftSpatialLocs}
\alias{getNoshiftSpatialLocs}
\title{getNoshiftSpatialLocs}
\usage{
getNoshiftSpatialLocs(object)
}
\arguments{
\item{object}{giotto object}
}
\description{
Coveret shiftData To NoShiftData
}
