% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GiottoToAnndata_2024.R
\name{Function giottoToAnnData}
\alias{Function giottoToAnnData}
\alias{giottoToAnnData}
\title{giottoToAnnData}
\usage{
giottoToAnnData(object = NULL, outpath = NULL, markerDF = NULL)
}
\arguments{
\item{object}{a giotto object}

\item{outpath}{the path to write h5ad}

\item{markerDF}{marker table to exprt}
}
\description{
Coveret Giotto To Anndata
}
