% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratToAnndata_2024.R
\name{Function SeuratToAnndata}
\alias{Function SeuratToAnndata}
\alias{SeuratToAnndata}
\title{SeuratToAnndata}
\usage{
SeuratToAnndata(
  object,
  outpath,
  assays,
  groups = NULL,
  reductions = NULL,
  markersDF = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{outpath}{output dir}

\item{assays}{vector of assay names to export}

\item{groups}{vector of groups to export, if null export all}

\item{reductions}{vector of reduction names to export, if null export all}

\item{markersDF}{a named list of marker df vars,need the colnames in c(gene, cluster, ...) order}
}
\description{
SeuratToAnndata
}
