% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratToAnndata_2024.R
\name{Gain_spatial}
\alias{Gain_spatial}
\title{Function used to parser spatial informs of different ST-seq}
\usage{
Gain_spatial(object, img.name, adata)
}
\arguments{
\item{object}{Seurat object}

\item{img.name}{name of the image}

\item{adata}{AnnData object}
}
\value{
a list
}
\description{
Function used to parser spatial informs of different ST-seq
}
