% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GiottoToAnndata_2023.R
\name{get_expression_values}
\alias{get_expression_values}
\title{Get expression values}
\usage{
get_expression_values(
  gobject,
  values = NULL,
  spat_unit = NULL,
  feat_type = NULL,
  output = c("exprObj", "matrix")
)
}
\arguments{
\item{gobject}{image object}

\item{values}{expression values to extract (e.g. "raw", "normalized", "scaled")}

\item{spat_unit}{spatial unit ‘cell’}

\item{feat_type}{feature type ‘rna’}

\item{output}{what object type to retrieve the expression as. Currently either
'matrix' for the matrix object contained in the exprObj or 'exprObj' (default) for
the exprObj itself are allowed.}
}
\value{
expression matrix
}
\description{
Function to get expression values from giotto object
}
\concept{expression accessor functions}
\concept{functions to get data from giotto object}
