% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GiottoToAnndata_2023.R
\name{set_default_feat_type}
\alias{set_default_feat_type}
\title{set_default_feat_type}
\usage{
set_default_feat_type(gobject, feat_type = NULL, spat_unit)
}
\arguments{
\item{gobject}{gobject}

\item{feat_type}{feature type ‘rna’}

\item{spat_unit}{spatial unit ‘cell’}
}
\description{
set_default_feat_type
}
