% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GiottoToAnndata_2024.R
\name{set_default_spat_unit}
\alias{set_default_spat_unit}
\title{set_default_spat_unit}
\usage{
set_default_spat_unit(gobject, spat_unit = NULL)
}
\arguments{
\item{gobject}{gobject}

\item{spat_unit}{spatial unit ‘cell’}
}
\description{
set_default_spat_unit
}
\keyword{internal}
