% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GiottoToAnndata_2023.R
\name{Function getNoshiftSpatialLocs}
\alias{Function getNoshiftSpatialLocs}
\alias{getNoshiftSpatialLocs}
\title{getNoshiftSpatialLocs}
\usage{
getNoshiftSpatialLocs(object, cell_meta_anno = NULL, spatial_locs = NULL)
}
\arguments{
\item{object}{giotto object}

\item{cell_meta_anno}{cell meta table}

\item{spatial_locs}{spatial location}
}
\description{
Coveret shiftData To NoShiftData
}
